Refactor object and property metadatas to remove extractor (#18)
